chore: refactor bitcoinrpc to envs
