+++ fedimint-core/src/lib.rs
@@ -28,7 +28,6 @@ pub mod admin_client;
 pub mod api;
 pub mod backup;
 pub mod bitcoin_migration;
-pub mod bitcoinrpc;
 pub mod cancellable;
 pub mod config;
 pub mod core;
