+++ modules/fedimint-wallet-common/src/config.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeMap;
 
 use bitcoin::Network;
-use fedimint_core::bitcoinrpc::BitcoinRpcConfig;
+use fedimint_core::envs::BitcoinRpcConfig;
 use fedimint_core::core::ModuleKind;
 use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::module::__reexports::serde_json;
